+++ gss/subgraph_isomorphism_test.cc
@@ -44,7 +44,7 @@ R"(1,
     {
         params.count_solutions = true;
         auto result = solve_homomorphism_problem(pattern, target, params);
-        CHECK(result.solution_count == long(4 * 3 * 2));
+        CHECK(result.solution_count == 4 * 3 * 2);
         CHECK(result.complete);
     }
 }
@@ -70,7 +70,7 @@ R"(1,2
     {
         params.count_solutions = true;
         auto result = solve_homomorphism_problem(pattern, target, params);
-        CHECK(result.solution_count == 4L);
+        CHECK(result.solution_count == 4);
         CHECK(result.complete);
     }
 }
@@ -97,7 +97,7 @@ R"(1,2
     {
         params.count_solutions = true;
         auto result = solve_homomorphism_problem(pattern, target, params);
-        CHECK(result.solution_count == 2L);
+        CHECK(result.solution_count == 2);
         CHECK(result.complete);
     }
 }
